add and remove segment
Reworked the areValid method for the segments. Had a go at the add/remove, but not sure. If I get the add segment to work, I break the legal test case for the one above. Added the totalLength method, did not yet write the proper implementation for isTraversable or getAllLocations