Added main method
Added the 4 requirements for the new main. Changed some toString representations so that for example location does not return the coordinates anymore, but just the address, it looks nicer I think. Had to do a for loop for the route toString because it kept referencing the hashCode when I tried to print it